cambio a el evento click en la tabla record
Se hizo un cambio a el evento click de la tabla en la vista Record ese era el error que tenia (.toString) ya corregido y algunas ediciones a textos.
