--- ExampleJDBC/src/main/java/co/edu/sena/examplejdbc/view/JFrameRecordModified.java
@@ -1,562 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package co.edu.sena.examplejdbc.view;
-
-import co.edu.sena.examplejdbc.controller.EmployeeController;
-import co.edu.sena.examplejdbc.controller.IEmployeeController;
-import co.edu.sena.examplejdbc.controller.IKeyController;
-import co.edu.sena.examplejdbc.controller.IRecordController;
-import co.edu.sena.examplejdbc.controller.KeyController;
-import co.edu.sena.examplejdbc.controller.RecordController;
-import co.edu.sena.examplejdbc.model.Employee;
-import co.edu.sena.examplejdbc.model.Key;
-import co.edu.sena.examplejdbc.model.Record;
-import co.edu.sena.examplejdbc.utils.MessageUtils;
-import java.awt.Color;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.List;
-import javax.swing.DefaultComboBoxModel;
-import javax.swing.JOptionPane;
-import javax.swing.table.DefaultTableModel;
-
-/**
- * date: 05/04/2025
- *
- * @author Esteban Colorado Vargas
- */
-public class JFrameRecordModified extends javax.swing.JFrame {
-
-    private final IRecordController recordController = new RecordController();
-    private final IEmployeeController employeeController = new EmployeeController();
-    private final IKeyController keyController = new KeyController();
-    int xMouse;
-    int yMouse;
-
-    /**
-     * Creates new form JFrameRecordModified
-     */
-    public JFrameRecordModified() {
-        initComponents();
-        fillTable();
-        fillCombos();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jPanelBackground = new javax.swing.JPanel();
-        jScrollPane1 = new javax.swing.JScrollPane();
-        jTable1 = new javax.swing.JTable();
-        jLabeltTitle1 = new javax.swing.JLabel();
-        jButtonCleart = new javax.swing.JButton();
-        jButtonDelete = new javax.swing.JButton();
-        jButtonUpdate = new javax.swing.JButton();
-        jButtonInsert = new javax.swing.JButton();
-        jComboBoxStatus = new javax.swing.JComboBox<>();
-        jLabelStatus = new javax.swing.JLabel();
-        jComboBoxKeyId = new javax.swing.JComboBox();
-        jLabelKey_Name = new javax.swing.JLabel();
-        timePickerEnd = new com.github.lgooddatepicker.components.TimePicker();
-        jLabelEnd_Time = new javax.swing.JLabel();
-        DatePickerDay = new com.github.lgooddatepicker.components.DatePicker();
-        jLabelDate_Record = new javax.swing.JLabel();
-        jComboBoxType_id = new javax.swing.JComboBox<>();
-        jLabelType_id = new javax.swing.JLabel();
-        timePickerStart = new com.github.lgooddatepicker.components.TimePicker();
-        jLabelStart_Time = new javax.swing.JLabel();
-        jTextFieldId = new javax.swing.JTextField();
-        jLabelId = new javax.swing.JLabel();
-        jLabelDescript = new javax.swing.JLabel();
-        jLabeltTitle = new javax.swing.JLabel();
-        jPanelCloseWindow = new javax.swing.JPanel();
-        jLabelCloseWindow = new javax.swing.JLabel();
-        jLabelBackground = new javax.swing.JLabel();
-        jPanelHeader = new javax.swing.JPanel();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setLocationByPlatform(true);
-        setUndecorated(true);
-        setResizable(false);
-
-        jPanelBackground.setBackground(new java.awt.Color(255, 255, 255));
-        jPanelBackground.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        jTable1.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
-        jTable1.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-
-            },
-            new String [] {
-
-            }
-        ));
-        jTable1.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                jTable1MouseClicked(evt);
-            }
-        });
-        jScrollPane1.setViewportView(jTable1);
-
-        jPanelBackground.add(jScrollPane1, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 350, 750, 130));
-
-        jLabeltTitle1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabeltTitle1.setText("HISTORIAL DE REGISTROS");
-        jLabeltTitle1.setFont(new java.awt.Font("Calibri", 1, 18)); // NOI18N
-        jLabeltTitle1.setForeground(new java.awt.Color(255, 255, 255));
-        jPanelBackground.add(jLabeltTitle1, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 320, 770, -1));
-
-        jButtonCleart.setBackground(new java.awt.Color(0, 204, 204));
-        jButtonCleart.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jButtonCleart.setForeground(new java.awt.Color(255, 255, 255));
-        jButtonCleart.setText("LIMPIAR");
-        jButtonCleart.setBorderPainted(false);
-        jButtonCleart.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
-        jButtonCleart.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButtonCleartActionPerformed(evt);
-            }
-        });
-        jPanelBackground.add(jButtonCleart, new org.netbeans.lib.awtextra.AbsoluteConstraints(560, 280, 120, -1));
-
-        jButtonDelete.setBackground(new java.awt.Color(255, 51, 51));
-        jButtonDelete.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jButtonDelete.setForeground(new java.awt.Color(255, 255, 255));
-        jButtonDelete.setText("ELIMINAR");
-        jButtonDelete.setBorderPainted(false);
-        jButtonDelete.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
-        jButtonDelete.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButtonDeleteActionPerformed(evt);
-            }
-        });
-        jPanelBackground.add(jButtonDelete, new org.netbeans.lib.awtextra.AbsoluteConstraints(410, 280, 120, -1));
-
-        jButtonUpdate.setBackground(new java.awt.Color(102, 102, 255));
-        jButtonUpdate.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jButtonUpdate.setForeground(new java.awt.Color(255, 255, 255));
-        jButtonUpdate.setText("ACTUALIZAR");
-        jButtonUpdate.setBorderPainted(false);
-        jButtonUpdate.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
-        jButtonUpdate.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButtonUpdateActionPerformed(evt);
-            }
-        });
-        jPanelBackground.add(jButtonUpdate, new org.netbeans.lib.awtextra.AbsoluteConstraints(260, 280, 120, -1));
-
-        jButtonInsert.setBackground(new java.awt.Color(0, 255, 0));
-        jButtonInsert.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jButtonInsert.setForeground(new java.awt.Color(255, 255, 255));
-        jButtonInsert.setText("INSERTAR");
-        jButtonInsert.setBorderPainted(false);
-        jButtonInsert.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
-        jButtonInsert.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButtonInsertActionPerformed(evt);
-            }
-        });
-        jPanelBackground.add(jButtonInsert, new org.netbeans.lib.awtextra.AbsoluteConstraints(110, 280, 120, -1));
-
-        jComboBoxStatus.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "ENTREGADO", "PENDIENTE", "CANCELADO" }));
-        jPanelBackground.add(jComboBoxStatus, new org.netbeans.lib.awtextra.AbsoluteConstraints(190, 230, 170, -1));
-
-        jLabelStatus.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jLabelStatus.setForeground(new java.awt.Color(255, 255, 255));
-        jLabelStatus.setText("ESTADO:");
-        jPanelBackground.add(jLabelStatus, new org.netbeans.lib.awtextra.AbsoluteConstraints(50, 230, -1, -1));
-
-        jPanelBackground.add(jComboBoxKeyId, new org.netbeans.lib.awtextra.AbsoluteConstraints(560, 190, 180, -1));
-
-        jLabelKey_Name.setText("NOMBRE DE LLAVE:");
-        jLabelKey_Name.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jLabelKey_Name.setForeground(new java.awt.Color(255, 255, 255));
-        jPanelBackground.add(jLabelKey_Name, new org.netbeans.lib.awtextra.AbsoluteConstraints(390, 190, -1, -1));
-        jPanelBackground.add(timePickerEnd, new org.netbeans.lib.awtextra.AbsoluteConstraints(560, 150, 180, -1));
-
-        jLabelEnd_Time.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jLabelEnd_Time.setForeground(new java.awt.Color(255, 255, 255));
-        jLabelEnd_Time.setText("HORA DE FINALIZACION:");
-        jPanelBackground.add(jLabelEnd_Time, new org.netbeans.lib.awtextra.AbsoluteConstraints(390, 150, -1, -1));
-        jPanelBackground.add(DatePickerDay, new org.netbeans.lib.awtextra.AbsoluteConstraints(560, 110, 180, -1));
-
-        jLabelDate_Record.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jLabelDate_Record.setForeground(new java.awt.Color(255, 255, 255));
-        jLabelDate_Record.setText("FECHA DE REGISTRO:");
-        jPanelBackground.add(jLabelDate_Record, new org.netbeans.lib.awtextra.AbsoluteConstraints(390, 110, -1, -1));
-
-        jPanelBackground.add(jComboBoxType_id, new org.netbeans.lib.awtextra.AbsoluteConstraints(190, 190, 170, -1));
-
-        jLabelType_id.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jLabelType_id.setForeground(new java.awt.Color(255, 255, 255));
-        jLabelType_id.setText("NOMBRE EMPLEADO:");
-        jPanelBackground.add(jLabelType_id, new org.netbeans.lib.awtextra.AbsoluteConstraints(50, 190, -1, -1));
-        jPanelBackground.add(timePickerStart, new org.netbeans.lib.awtextra.AbsoluteConstraints(190, 150, 170, -1));
-
-        jLabelStart_Time.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jLabelStart_Time.setForeground(new java.awt.Color(255, 255, 255));
-        jLabelStart_Time.setText("HORA DE INICIO:");
-        jPanelBackground.add(jLabelStart_Time, new org.netbeans.lib.awtextra.AbsoluteConstraints(50, 150, -1, -1));
-
-        jTextFieldId.setFont(new java.awt.Font("Segoe UI", 1, 12)); // NOI18N
-        jPanelBackground.add(jTextFieldId, new org.netbeans.lib.awtextra.AbsoluteConstraints(190, 110, 170, -1));
-
-        jLabelId.setFont(new java.awt.Font("Roboto Condensed", 1, 14)); // NOI18N
-        jLabelId.setForeground(new java.awt.Color(255, 255, 255));
-        jLabelId.setText("ID:");
-        jPanelBackground.add(jLabelId, new org.netbeans.lib.awtextra.AbsoluteConstraints(50, 110, -1, -1));
-
-        jLabelDescript.setFont(new java.awt.Font("Roboto Condensed", 0, 14)); // NOI18N
-        jLabelDescript.setForeground(new java.awt.Color(255, 255, 255));
-        jLabelDescript.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabelDescript.setText("Aqui  puedes insertar, actualizar, eliminar y consultar todo lo referente a la tabla Registro.");
-        jPanelBackground.add(jLabelDescript, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 80, 770, -1));
-
-        jLabeltTitle.setFont(new java.awt.Font("Roboto Condensed", 1, 20)); // NOI18N
-        jLabeltTitle.setForeground(new java.awt.Color(255, 255, 255));
-        jLabeltTitle.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabeltTitle.setText("CREAR REGISTRO");
-        jPanelBackground.add(jLabeltTitle, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 50, 770, -1));
-
-        jPanelCloseWindow.setBackground(new java.awt.Color(255, 255, 255));
-
-        jLabelCloseWindow.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabelCloseWindow.setText("x");
-        jLabelCloseWindow.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
-        jLabelCloseWindow.setFont(new java.awt.Font("Roboto", 0, 24)); // NOI18N
-        jLabelCloseWindow.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                jLabelCloseWindowMouseClicked(evt);
-            }
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                jLabelCloseWindowMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                jLabelCloseWindowMouseExited(evt);
-            }
-        });
-
-        javax.swing.GroupLayout jPanelCloseWindowLayout = new javax.swing.GroupLayout(jPanelCloseWindow);
-        jPanelCloseWindow.setLayout(jPanelCloseWindowLayout);
-        jPanelCloseWindowLayout.setHorizontalGroup(
-            jPanelCloseWindowLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jLabelCloseWindow, javax.swing.GroupLayout.DEFAULT_SIZE, 50, Short.MAX_VALUE)
-        );
-        jPanelCloseWindowLayout.setVerticalGroup(
-            jPanelCloseWindowLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jLabelCloseWindow, javax.swing.GroupLayout.DEFAULT_SIZE, 50, Short.MAX_VALUE)
-        );
-
-        jPanelBackground.add(jPanelCloseWindow, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 50, 50));
-
-        jLabelBackground.setIcon(new javax.swing.ImageIcon(getClass().getResource("/co/edu/sena/examplejdbc/view/background para pestañas.png"))); // NOI18N
-        jPanelBackground.add(jLabelBackground, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 770, 490));
-
-        jPanelHeader.addMouseMotionListener(new java.awt.event.MouseMotionAdapter() {
-            public void mouseDragged(java.awt.event.MouseEvent evt) {
-                jPanelHeaderMouseDragged(evt);
-            }
-        });
-        jPanelHeader.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mousePressed(java.awt.event.MouseEvent evt) {
-                jPanelHeaderMousePressed(evt);
-            }
-        });
-
-        javax.swing.GroupLayout jPanelHeaderLayout = new javax.swing.GroupLayout(jPanelHeader);
-        jPanelHeader.setLayout(jPanelHeaderLayout);
-        jPanelHeaderLayout.setHorizontalGroup(
-            jPanelHeaderLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 770, Short.MAX_VALUE)
-        );
-        jPanelHeaderLayout.setVerticalGroup(
-            jPanelHeaderLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 50, Short.MAX_VALUE)
-        );
-
-        jPanelBackground.add(jPanelHeader, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 770, 50));
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanelBackground, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanelBackground, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-
-        pack();
-        setLocationRelativeTo(null);
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void jTable1MouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jTable1MouseClicked
-        // SELECCIONAR REGISTROS DE LA TABLA
-        int rowSelected = jTable1.getSelectedRow();
-        if (rowSelected != -1) {
-            int idSelected = Integer.parseInt(jTable1.getValueAt(rowSelected, 0).toString());
-            try {
-                Record record = recordController.findById(idSelected);
-                jTextFieldId.setText(String.valueOf(idSelected));
-                DatePickerDay.setDate(LocalDate.parse(record.getDate_Record()));
-                timePickerStart.setText(record.getStart_time());
-                timePickerEnd.setText(record.getEnd_time());
-                jComboBoxType_id.getModel().setSelectedItem(record.getEmployeeId());
-                jComboBoxKeyId.getModel().setSelectedItem(record.getKey().toString());
-                jComboBoxStatus.setSelectedItem(record.getStatus());
-            } catch (Exception e) {
-                MessageUtils.showErrorMessage("ERROR al usar el evento de click"
-                        + e.getMessage());
-            }
-        }
-        jButtonInsert.setEnabled(false);
-        jButtonDelete.setEnabled(true);
-        jButtonUpdate.setEnabled(true);
-    }//GEN-LAST:event_jTable1MouseClicked
-
-    private void jButtonCleartActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButtonCleartActionPerformed
-        // BOTON LIMPIAR
-        clear();
-    }//GEN-LAST:event_jButtonCleartActionPerformed
-
-    private void jButtonDeleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButtonDeleteActionPerformed
-        // BOTON PARA ELIMINAR
-        try {
-            int option = JOptionPane.showConfirmDialog(rootPane, "Estas seguro de eliminar el empleado?",
-                    "CONFIRMAR", JOptionPane.YES_NO_OPTION);
-            if (option == JOptionPane.YES_OPTION) {
-                recordController.delete(Integer.parseInt(jTextFieldId.getText()));
-                MessageUtils.showInfoMessage("Registro eliminado correctamente...");
-                fillTable();
-            }
-            clear();
-            insertDefaultDate();
-        } catch (Exception e) {
-            MessageUtils.showErrorMessage("HUBO UN ERROR al eliminar..." + e.getMessage());
-        }
-    }//GEN-LAST:event_jButtonDeleteActionPerformed
-
-    private void jButtonUpdateActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButtonUpdateActionPerformed
-        // ACTULIZAR
-       try {
-           Record record = new Record();
-           record.setId(Integer.parseInt(jTextFieldId.getText()));
-           record.setDate_Record(DatePickerDay.getDate().toString());
-           record.setStart_time(timePickerStart.getTime().toString());
-           record.setEnd_time(timePickerEnd.getTime() != null ? timePickerEnd.getTime().toString()
-                   : null);
-           record.setEmployeeId((Employee) jComboBoxType_id.getSelectedItem());
-           Key selectedKey = (Key) jComboBoxKeyId.getSelectedItem();
-           record.setKey(selectedKey);
-           record.setStatus(jComboBoxStatus.getSelectedItem().toString());
-           recordController.update(record);
-           MessageUtils.showInfoMessage("Registro actualizado exitosamente");
-           clear();
-           fillTable();
-       } catch (Exception e) {
-           MessageUtils.showErrorMessage(e.getMessage());
-        }
-    }//GEN-LAST:event_jButtonUpdateActionPerformed
-
-    private void jButtonInsertActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButtonInsertActionPerformed
-        // BOTON PARA INSERTAR
-        try {
-            Record record = new Record();
-            record.setDate_Record(DatePickerDay.getDate().toString());
-            record.setStart_time(timePickerStart.getTimeStringOrEmptyString());
-            record.setEnd_time(null);
-            record.setEmployeeId((Employee) jComboBoxType_id.getSelectedItem());
-            Key key = (Key) jComboBoxKeyId.getSelectedItem();
-            record.setKey(key);
-            record.setStatus(jComboBoxStatus.getSelectedItem().toString());
-            recordController.insert(record);
-            MessageUtils.showInfoMessage("El registro se ha insertado correctamente...");
-            fillTable();
-            clear();
-            insertDefaultDate();
-        } catch (Exception e) {
-            MessageUtils.showErrorMessage("Ha ocurrido un error al insertar el usuario..."
-                    + e.getMessage());
-        }
-    }//GEN-LAST:event_jButtonInsertActionPerformed
-
-    private void jPanelHeaderMousePressed(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jPanelHeaderMousePressed
-        // EVENTO COMPLEMETNARIO PARA MOVER LA PESTAÑA
-        xMouse = evt.getX();
-        yMouse = evt.getY();
-    }//GEN-LAST:event_jPanelHeaderMousePressed
-
-    private void jPanelHeaderMouseDragged(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jPanelHeaderMouseDragged
-        // EVENTO PARA MOVER LA PESTAÑA
-        int xPosition = evt.getXOnScreen();
-        int yPosition = evt.getYOnScreen();
-
-        this.setLocation(xPosition - xMouse, yPosition - yMouse);
-    }//GEN-LAST:event_jPanelHeaderMouseDragged
-
-    private void jLabelCloseWindowMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jLabelCloseWindowMouseClicked
-        // EVENTO PARA CERRAR EL PROGRAMA
-        int option = JOptionPane.showConfirmDialog(rootPane,
-                "Estas seguro de salir?", "CONFIRMAR", JOptionPane.YES_NO_OPTION);
-
-        if (option == JOptionPane.YES_OPTION) {
-            new JFrameMenu().setVisible(true);
-            dispose();
-        }
-    }//GEN-LAST:event_jLabelCloseWindowMouseClicked
-
-    private void jLabelCloseWindowMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jLabelCloseWindowMouseEntered
-        // BOTON PARA EFECTO HOVER
-        jPanelCloseWindow.setBackground(Color.red);
-    }//GEN-LAST:event_jLabelCloseWindowMouseEntered
-
-    private void jLabelCloseWindowMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jLabelCloseWindowMouseExited
-        // BOTON PARA EFECTO HOVER DE SALIDA
-        jPanelCloseWindow.setBackground(Color.white);
-    }//GEN-LAST:event_jLabelCloseWindowMouseExited
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(JFrameRecordModified.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(JFrameRecordModified.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(JFrameRecordModified.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(JFrameRecordModified.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new JFrameRecordModified().setVisible(true);
-            }
-        });
-    }
-
-    public void clear() {
-        jTextFieldId.setText("");
-        DatePickerDay.setText("");
-        timePickerStart.setText("");
-        timePickerEnd.setText("");
-        jComboBoxType_id.setSelectedIndex(0);
-        jComboBoxKeyId.setSelectedIndex(0);
-        jComboBoxStatus.setSelectedIndex(0);
-        jButtonInsert.setEnabled(true);
-        jButtonUpdate.setEnabled(false);
-        jButtonDelete.setEnabled(false);
-        insertDefaultDate();
-    }
-
-    public void fillTable() {
-        try {
-            DefaultTableModel model = new DefaultTableModel();
-            jTable1.setModel(model);
-            model.addColumn("ID");
-            model.addColumn("Fecha");
-            model.addColumn("Hora inicio");
-            model.addColumn("Hora finalizacion");
-            model.addColumn("Funcionario");
-            model.addColumn("LLave");
-            model.addColumn("Estado");
-
-            String[] rows = new String[7];
-            List<Record> results = recordController.findAll();
-            for (Record result : results) {
-                rows[0] = String.valueOf(result.getId());
-                rows[1] = result.getDate_Record();
-                rows[2] = result.getStart_time();
-                rows[3] = result.getEnd_time();
-                rows[4] = result.getEmployeeId().getFullName();
-                rows[5] = result.getKey().getName();
-                rows[6] = result.getStatus();
-                model.addRow(rows);
-            }
-            jTextFieldId.setEnabled(false);
-            jButtonInsert.setEnabled(true);
-            jButtonUpdate.setEnabled(false);
-            jButtonDelete.setEnabled(false);
-        } catch (Exception e) {
-        }
-    }
-
-    public void fillCombos() {
-        try {
-            List<Employee> employees = employeeController.findAll();
-            DefaultComboBoxModel modelEmployee = new DefaultComboBoxModel();
-            jComboBoxType_id.setModel(modelEmployee);
-            modelEmployee.addAll(employees);
-
-            List<Key> allIdKeys = keyController.findAll();
-            DefaultComboBoxModel modelKeys = new DefaultComboBoxModel();
-            jComboBoxKeyId.setModel(modelKeys);
-            modelKeys.addAll(allIdKeys);
-            jComboBoxKeyId.setSelectedIndex(0);
-            jComboBoxType_id.setSelectedIndex(0);
-
-        } catch (Exception e) {
-            MessageUtils.showErrorMessage(e.getMessage());
-        }
-    }
-
-    public void insertDefaultDate() {
-        DatePickerDay.setDate(LocalDate.now());
-        timePickerStart.setTime(LocalTime.now());
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private com.github.lgooddatepicker.components.DatePicker DatePickerDay;
-    private javax.swing.JButton jButtonCleart;
-    private javax.swing.JButton jButtonDelete;
-    private javax.swing.JButton jButtonInsert;
-    private javax.swing.JButton jButtonUpdate;
-    private javax.swing.JComboBox jComboBoxKeyId;
-    private javax.swing.JComboBox<String> jComboBoxStatus;
-    private javax.swing.JComboBox<String> jComboBoxType_id;
-    private javax.swing.JLabel jLabelBackground;
-    private javax.swing.JLabel jLabelCloseWindow;
-    private javax.swing.JLabel jLabelDate_Record;
-    private javax.swing.JLabel jLabelDescript;
-    private javax.swing.JLabel jLabelEnd_Time;
-    private javax.swing.JLabel jLabelId;
-    private javax.swing.JLabel jLabelKey_Name;
-    private javax.swing.JLabel jLabelStart_Time;
-    private javax.swing.JLabel jLabelStatus;
-    private javax.swing.JLabel jLabelType_id;
-    private javax.swing.JLabel jLabeltTitle;
-    private javax.swing.JLabel jLabeltTitle1;
-    private javax.swing.JPanel jPanelBackground;
-    private javax.swing.JPanel jPanelCloseWindow;
-    private javax.swing.JPanel jPanelHeader;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JTable jTable1;
-    private javax.swing.JTextField jTextFieldId;
-    private com.github.lgooddatepicker.components.TimePicker timePickerEnd;
-    private com.github.lgooddatepicker.components.TimePicker timePickerStart;
-    // End of variables declaration//GEN-END:variables
-}
